+++ src/main/types.ts
@@ -428,6 +428,7 @@ type CloudObject = {
 interface ICloudClient {
   list: () => Promise<CloudObject[]>;
   delete: (key: string) => Promise<void>;
+  deleteVideo: (videoName: string) => Promise<void>;
 }
 
 interface IBrowserWindow {